refactor: added validate email and password to utils folder
