+++ src/utils/authenticationUtils/index.js
@@ -1,0 +1,2 @@
+export * from './validateEmail'
+export * from './validatePass'
