+++ src/utils/authenticationUtils/validatePass.js
@@ -1,0 +1,5 @@
+const validatePass = pass => {
+    return pass.length >= 8
+}
+
+export { validatePass }
